+++ buttons/buttons_search_by_distance.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//09/06/23
+//20/09/23
 
 include('..\\helpers\\buttons_xxx.js');
 include('..\\helpers\\helpers_xxx_properties.js');
@@ -36,6 +36,7 @@ delete newButtonsProperties.bHarmonicMixDoublePass;
 delete newButtonsProperties.bAscii;
 delete newButtonsProperties.bAdvTitle;
 delete newButtonsProperties.sortBias;
+delete newButtonsProperties.regionFilter;
 setProperties(newButtonsProperties, prefix, 0); //This sets all the panel properties at once
 {
 	const properties = getPropertiesPairs(newButtonsProperties, prefix, 0);